requirements.txt
black==22.1.0
click==8.0.4
Flask==2.0.3
Flask-Login==0.5.0
Flask-Reuploaded==1.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.22.3
pandas==1.4.1
pathspec==0.9.0
platformdirs==2.5.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.32
tomli==2.0.1
typing-extensions==4.1.1
Werkzeug==2.0.3
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
